pyproject.toml
[project]
name = "longweekend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-alembic>=3.2.0",
    "flask-sqlalchemy>=3.1.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
